CAMEL-18456: converted camel-atom to camel-test-infra-jetty
